simple calculator GUI
This app is a simple GUE for a calculator.
I first designed the GUA using Qty Designer and then converted it to Python code.
The calculator does not work and is just a GUA.
Maybe in the future I will work on the logic behind the calculator, although it is a challenging task.